+++ sdk/constants/stats.ts
@@ -1,0 +1,7 @@
+export const DEFAULT_LEADERBOARD_DATA = {
+	top: [],
+	bottom: [],
+	wallet: [],
+	search: [],
+	all: [],
+};
