+++ sdk/index.ts
@@ -5,7 +5,6 @@ import ExchangeService from './services/exchange';
 import FuturesService from './services/futures';
 import KwentaTokenService from './services/kwentaToken';
 import PricesService from './services/prices';
-import StakingService from './services/staking';
 import SynthsService from './services/synths';
 import SystemService from './services/system';
 import TransactionsService from './services/transactions';
@@ -19,7 +18,6 @@ export default class KwentaSDK {
 	public transactions: TransactionsService;
 	public kwentaToken: KwentaTokenService;
 	public prices: PricesService;
-	public staking: StakingService;
 	public system: SystemService;
 
 	constructor(context: IContext) {
@@ -30,7 +28,6 @@ export default class KwentaSDK {
 		this.synths = new SynthsService(this);
 		this.transactions = new TransactionsService(this);
 		this.kwentaToken = new KwentaTokenService(this);
-		this.staking = new StakingService(this);
 		this.system = new SystemService(this);
 	}
